Refactor Day01.kt and increase its functionality
The file Day01.kt was renamed and moved to a new directory, part of the structure, to unify the project's organization. Also, new functionality was added. The solution for 'part1' and 'part2' methods were rewritten to calculate calibration values for each line now. Two helper functions 'getCalibrationValueForEachLine' and 'convertLetterDigitToInt' were introduced to improve readability and maintainability of the code. This refactoring allows handling both number and digit-letter mixed inputs.