Add Day06 game with input file and calculation functions
Added Day06 game code and corresponding input file. The game includes functions for the calculation of optimal game moves, present as a multiplication of passing game indices. These computations are separate for the first and the second part of the game. Inputs are read directly from the Day06_input.txt file.